requirements.txt

Flask==2.3.2
openai-whisper==20230314